requirements.txt
Pillow==9.4.0
python-decouple==3.7
pyTelegramBotAPI==4.10